restructer dashboard layout so links can have required roles similar to the main navbar
