+++ src/pages/dashboard/bills.tsx
@@ -1,0 +1,12 @@
+/**
+ * Main bills page. Will only show bills tied to the user.
+ */
+const BillsDashboardPage = () => {
+    return (
+        <main>
+
+        </main>
+    );
+};
+
+export default BillsDashboardPage;
