+++ src/pages/payments/index.tsx
@@ -1,0 +1,12 @@
+/**
+ * Main payments page. Will show all patients payments.
+ */
+const PaymentsPage = () => {
+    return (
+        <main>
+
+        </main>
+    );
+};
+
+export default PaymentsPage;
